Add logtailer server port to PeerAttrsPB for BLS peers
Summary:
The logtailer server port will be set and persisted in topology config and config change events. This port is used to identify already existing peers that are not backed by a database so that we do not end up creating duplicate services for them.

Test Plan:
Tested with plugin in fbcode

Reviewers: arahut, yichenshen

Subscribers:

Tasks:

Tags: