[proxy] [perf] [3/n] avoid reading ops in leader for a proxied request
Summary:
When the leader has to send a request to a 'proxied' peer, it sends a request to
the proxy host which in turn reconstitutes the request for the peer.

The request sent by the leader to the proxy host contains all the ReplicateMsg
ops (with OpId populated). This essentially means that the leader has to read
the ops from LogCache (potentially hitting disk) and loop over to build multiple
ops in the batch (by discarding the payload). This is not needed since the proxy
host is going to reconstitute all the ops (with payload) again by reading from
its local LogCache. The leader has the simply send the 'next_index' that has to
be sent to the proxied peer and the proxy host can use this to reconstitute the
ops.

This PR tries to implement the same.
1. PeerMessageQueue::RequestForPeer() on leader identifies that this particular
peer is a 'proxied peer'. It then creates a single ReplicateMsg containing the
OpId with the 'next_index' for the peer (and a dummy term '0'). This does not
change the raft protocol since the leader is still the one deciding the
'next_index' for the proxied peer.
2. RaftConsensus::HandleProxyRequest() on the proxy host reconstitutes the
messages to be proxied based on the 'next_index' set in the leader's request.
3. The proxied peer is also responsible now for populating 'preceding_id' field
in the AppendEntries() RPC request sent to the proxied peer. "preceding_id"
is the OpId immediately before the 'next_index' in the leader's log. This might
seem like a divergence from raft protocol. "preceding_id" will now be the
(previousLogIndex, previousLogTerm) in the proxied peer's log rather than the
leader's log. However, LMP property ensures that the proxied peer's and the
leader's log are exactly the same. Hence this is safe by induction.

Overall benefit:
1. Avoild doing unecessary loops on leader to build the entire batch of Ops
which will be discarded anyways.
2. This is much more useful when the proxied peer is lagging the the ops have
fallen out of log cache. This avoids doing disk reads which can be common when
there there are a lot of proxied peer's in the ring.

Test Plan:
1. local tp2 checkout and run existing functional tests for proxy
2. if feasible, will do perf tests

Reviewers:

Subscribers:

Tasks:

Tags: