Add support for persisting variables
Summary:
When a configuration/variable change needs to be persisted, we will now write the variable/value to a new "persistent_vars" proto file. This file will be read during the `Init()` phase of consensus. Actions can be taken based on the values read from the file.

As an example, we make it so that whenever failure detector is disabled/enabled, we will write this to the proto file. When consensus is Init'd, we will read this value from the file and stop failure detector if the value read is True.

Test Plan:
Tested as part of integration tests in fbcode

Reviewers: vinaybhat, arahut, abhinavsharma, mpercy, chili

Subscribers:

Tasks:

Tags: