[peer-management] [FR] [1/n]  perf enhancements in peer management for FR
Summary:
While FlexiRaft adds the flexibility of having a smaller data commit
quorum (for faster commits) and a wider leader election quorum (for more
failover options), the underlying data structures are still shared by all
the peers. This creates contention points as the number of peers in the
ring grows.

This PR is a first step in splitting the underlying data structres which are
responsible for managing different peers and queues. The rationale is
that this will reduce contention in these shared data structures.

There are two major changes in this PR

1. PeerManager to have two different lists of peers:
PeerManager manages a list of peers. Whenever an operation needs to be
replicated, RaftConsensus::Replicate() calls into
PeerManager::SignalRequest() to start this process.
PeerManager::SignalRequest() is also called when the commit index moves
forward (RaftConsensus::NotifyCommitIndex(...)). This PR makes PeerManager
manage two different lists - one for peers that are in data commit
quorum and the other for everything else. PeerManager::SignalRequest() tries to
build the request first for all the members in data commit quorum and
follows it up by building the same for other peers. This ensures that
any lagging peer that is not in the critical path of committing a trx
will not block out commits. Note that the determination of a peer being a
member of data commit quorum is tightly coupled with the concept of
'leader region' in SINGLE_REGION_DYNAMIC mode of FR. Hence this optimization
is possible only in this mode.

2. Enhancements to "request_pending" tracker of the Peer:
For every peer, the leader  tracks if a requent is pending i.e a request is in
flight for this particular peer. There can only be one outstanding request for
a peer. This PR changes the semantics of 'request_pending' to mean 'request
is being built' (instead of 'request is in flight'). Building a request
could be expensive for a lgging peer as it need to di disk IO. This
could block threads which are holding onto RaftConsensus::lock_ and
PeerManager::lock_, thus blocking out the commit process completely.
This PR makes it so that when a request is being built (i.e
Peer::SendNextRequest(...) calls into PeerMessageQueue::RequestForPeer()), we
return early when attempting to build another request for the same peer.
The rationale is that if there are multiple calls into
Peer::SendNextRequest() (from different callsites), then there is no
need to build duplicate requests for this particular peer. Any
additional opeartions can always be shipped in a subsequent batch. To
make this check lock-free, "request_pending_" is made an atomic.
This ensures that  RaftConsensus::lock_ and PeerManager::lock_ are not locked
for an extended duration

Future enhancements:
1. It is not clear why the advancement of commit index (on the leader)
should trigger a round of PeerManager::SignalRequest(). In a subsequent
PR, we can try avoiding calling PeerManager::SignalRequest() (from
RaftConsensus::NotifyCommitIndex()). We can even make it so that we only build
requests for peers in data commit quorum - the rest of the peers can get
requests when replicating the next operation OR the next heartbeat.
2. PeerMessageQueue() is another data structure which could use some
enhancements to better support FR topologies.

Test Plan:
1. tests in fbcode
2. pending perf tests

Reviewers:

Subscribers:

Tasks:

Tags: