Add log line for sync vs async leader transfer
Summary:

When transfering leadership there are 2 modes, a sync mode that'll start
the transfer immediately (wakes up target), or a async mode that'll wait
for the target to catch up before proceeding.

Add a log line so we can trace which happened.

Test Plan:

Logging change. Compile here for now. Quick to verify when integrated
with the plugin.

Reviewers: anirban-fb, vinaybhat

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>