Add a way to cancel TransferLeadership and ask for current state
Summary:

Expose an API in raft_consensus to allow consumers to cancel a
leadership transfer.

The idea is that when an external party times out waiting for the no-op
event after TransferLeadership, it can call into kudu to cancel the
Transfer and also get some state on whether we can treat the transfer as
void (never happened).

Since we only allow a single transfer to happen at any time, there's
only one transfer that we can cancel. The cancellation acts on the last
requested transfer (ending transfer is idempotent and can be called even
if there's no transfer ongoing), and returns OK if the last requested
transfer has not requested an election on an external node.

The cancel API is intentionally designed to be called even after
raft_consensus thinks it's out of a transfer period. This is so that if
raft_consensus times out internally before a external caller times out,
we can still extract information about what happened in the last
transfer.

Test Plan:

Just make sure it builds here, test it out in fbcode.

Reviewers: anirban-fb, bhatvinay

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>