Fix chained election flag
Summary:

Previously we assumed that anytime we get context when requested to
start an election, it's a chained election.

However, we wanted to pipe a context during a LMP to record downtime,
but that's a direct election (1 term bump).

Instead, check if the current leader is the same as the source in the
context (if present) to determine if this election is a chained election
triggered by a previous one (like leader => LBU => leader).

Test Plan:

Bring this into fbcode and check that the timestamps are right

Reviewers: anirban-fb, bhatvinay

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>