Record chained start time for dead promotions
Summary:

We want metrics on how long the ring is unavailable for writes. To
estimate that for dead promotions, we take the point of 1 heartbeat
after the start point of the failure detector's start time.

This should be a good proxy for the time of the first heartbeat failure,
and the leader should have failed somewhere between the failure
detector's start time and this time.

Test Plan:

Test in fbcode

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>