Fix the dead promotion time
Summary:

Previously, this recorded it as a monotime and converted it directly
into a system_clock timepoint. This isn't valid since monotime just
guarantees monotonic increase from a unspecified reference point, not
from epoch. Converting it to a timepoint yields a timestamp in the
1970s since the absolute valid of monotime is very low.

Fix this by using system_clock across the board and removing the
misleading timepoint from monotime function previously added.

Test Plan:

Build in fbcode and run dead promotions. The metric should be correct
when it's recorded directly from the ElectionContext created in the
failure detector

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>