Funnel the persistent_vars object into ConsensusQueue
Summary:

ConsensusQueue is what is building base Raft requests today. To be able
to stamp these with a ring token ConsensusQueue must be able to access
that token.

persistent_vars is available in RaftConsensus but ConsensusQueue is
abstracted away from RaftConsensus via an Observer interface.

Instead, persistent_vars is already a shared ptr, and the manager uses a
cache model that'll always return a loaded object if present. SO we get
ConsensusQueue to read from it too, kind of like a global library.

Test Plan:

We'll need to be actually using it to test it so just build on this
commit.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>