Get relevant RPCs to stamp the RPC tokens
Summary:

Add the raft_rpc_token field to the protobuf struct whenever we have a
token.

As of this commit no enforcement is done yet, we're only populating the
field.

I added it to only 3 RPCs. These are all initiated by consensus (or by
the plugin). I left anything that is meant to be initiated externally
alone, since this RPC is supposed to protect against other bad nodes,
not DBAs.

Added token:

- UpdateConsensus(ConsensusRequestPB)
  - This is the base consensus RPC for replicastion
- RequestConsensusVote(VoteRequestPB) returns (VoteResponsePB);
  - This is used during elections and will cause term bumps if it's not
    a pre-election
- RunLeaderElection(RunLeaderElectionRequestPB)
  - This is used during live promotions and is triggered from the
    plugin programmatically

The ones below are not stamped with a token, an are meant to be external RPCs.
A few of them, like ChangeConfig, eventually ends up firing an internal RPC like
UpdateConsensus.

- ChangeConfig(ChangeConfigRequestPB)
- BulkChangeConfig(BulkChangeConfigRequestPB)
- UnsafeChangeConfig(UnsafeChangeConfigRequestPB)
- ChangeProxyTopology(ChangeProxyTopologyRequestPB)
- GetNodeInstance(GetNodeInstanceRequestPB)
- LeaderStepDown(LeaderStepDownRequestPB)
- GetLastOpId(GetLastOpIdRequestPB)
- GetConsensusState(GetConsensusStateRequestPB)

Test Plan:

The next commit will add logging/enforcement, at which point we could
being testing proper.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>