Prevent RPC token from being changed when we're enforcing
Summary:

We should never change the RPC token while we're enforcing because it'll
immediately start breaking RPCs. As such we only allow RPC tokens to be
changed when enforcement is off.

Test Plan:

This will be tested in fbcode tests where we make sure token overrides
fail while we're enforcing.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>