Add RPC tokens to responses
Summary:

We previously rejected incoming RPCs if the token doesn't match our own
token.

Now we add this token to responses as well. We add it for a normal Raft
update and a vote request.

The rationale is that even if we reject incoming rpcs, we can still send
a RPC to a bad instance. That bad instance might not be enforcing rpc
tokens, and will give us a bad reply. For example, if we're enforcing
tokens but the peer is not, we can send a RPC and the peer can reply
with a higher term.

In this case, even though the peer does not have the right RPC, it will
still cause instability on our end simply by replying. We aim to prevent
that by ignoring responses that don't have the right token either.

Test Plan:

This implements one side of things. Compile for now and we can test when
we implement enforcement.

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>