Add a counter to check RPC mismatches
Summary:

Mismatches will bump the counter even if we're not enforcing.

Proxies are a bit tricky, and currently they bump counters on both the
proxy and the origin. This is because at the consensus_peers layer we
don't differentiate between proxies and origins. I still think it's ok
because we are aiming to get the counter totally to 0. Any value above 1
is bad anyway.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
Signed-off-by: Yichen <yichenshen@fb.com>