Fix reporting and clearing of errors from Proxies
Summary:  Errors added to the ErrorManager by a Proxy cannot be cleared because the entity, which is the host:port of the target instance, was set to UNKNOWN. When it came to clear the error, the proper entity was not found in the ErrorManager and the error never gets cleared. The reason entity is to UNKNOWN is because when a Proxy reports an error, it attempts to retrieve entity from its PeerManager. However, since the Proxy is not a leader, its PeerManager is empty, which causes us to default the entity name to UNKNOWN.

The fix is rather simple, we can simply get the peer's host:port from the active config.

Test Plan: Added tests in D39368699.

Reviewers:

Subscribers:

Tasks:

Tags: