Native Ingestion: Blocking iterator to read binlog events
Summary: In Native Ingestion, we need a way to iterate through new binlog events when they become available. This allows us to iterate through binlog events from an initial position and continue on and block when then there's no more events. It will resume when there are events.

Test Plan:
- Added unit tests.
- WIP on testing BinlogWrapper logic

Reviewed By: abhinav04sharma

Differential Revision: D50381018

fbshipit-source-id: 7f796db785418df956ad6e2677d30d44191d4247