Directly append new ops to the buffers
Summary:
Currently the code will reach into logcache to read new ops upon being signaled. This could cause contention on the lock in the log cache and incur extra CPU.

The new op is a shared pointer anyway. Append it directly to the buffer if possible.

There could be numerous cases where this fails (new ops are processed out of order, fills getting debounced). But in that case we can then read from the log cache. New ops are signaled in order, and in most cases should also be processed in order. Appends also happen under a peer lock, so it should not debounce itself (debouncing should be only from RPC needing handoffs).

Test Plan:
I canaried this on merlin with D47676500, which forces this function on and adds very verbose logging.

Start a workload and check that the buffer is being filled correctly and is being sent out correctly.

Reviewed By: abhinav04sharma

Differential Revision: D50245892

fbshipit-source-id: c7ec0c2bbc3e772f3f462046d3385d07963a35a5