Using replicate msg wrapper's codec instead of global codec in thrift binlog append
Summary:
We were using the global compression codec while appending in thrift
format. However, there is no guarantee that the global codec is the same as the
msg codec because we might receive multiple msgs from the leader in one batch
with different compression codecs and the global is always set to the latest
one. Now using the codec from the msg itself.

Test Plan:
Testing on Stage -1:

Before:
{P1064188657}

After:
{P1064191224}

Reviewed By: li-chi

Differential Revision: D52932254

fbshipit-source-id: 0d4c6d6c05f2a291cb325f0aac0091fb35af5c57