Remove unused variables in kudu/gutil/strings/escaping.cc
Summary:
LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Test Plan: Sandcastle

Reviewed By: bunnypak, dmm-fb

Differential Revision: D53011665

fbshipit-source-id: 187a64e0d0724ffcec366f1bbdd31b8c0bcd8f27