Add a method to attempt to serialize only the header of a RPC
Summary:
I've seen that for a large RPC call, connection.cc receives roughly 16KB at a time. This means that by the first receive call it should already have the full header (typically contained within the first 60 bytes or so).

From there, we should be able to get an early peek on both the size of the RPC, as well as the call. We can inject hooks with this, such a pausing the failure detector if we anticipate a large UpdateReplica call.

Test Plan: Mostly just refactoring here, will be tested in the next diff.

Reviewed By: alanliang

Differential Revision: D53676103

fbshipit-source-id: a0a4910a3088c16df1d375ad570dfba4a43b8ad2