Add protobuf generation for long call hooks
Summary:
We have a generator that generates MethodInfos for each RPC endpoint for ConsensusService in consensus.proto.

Currently, it registers `func` for the main function call, and `authz_method` for auth.

We add long_call_loading_hook and long_call_loaded_hook to this. By default they map to the default `LongCallLoading` and `LongCallLoaded` no-ops.

We can customize this by adding options to consensus proto, which I'll do in the next diff. The generator will automatically add abstract functions when we do this, and we'll override them in ConsensusServiceImpl.

Test Plan: I'll test this when the whole flow reaches RaftConsensus for a long UpdateReplica call.

Reviewed By: li-chi

Differential Revision: D53744060

fbshipit-source-id: ed5ba03161e01e17ab34887060b77533727ae430