Get services like ConsensusService to register hooks for long calls
Summary:
We hope to pause the failure detector in RaftConsensus when we get a very large UpdateConsensus call. These hooks lets RaftConsensus know that connections are reading a large RPC from the wire, as well as when it is done.

For this diff we connect it till ServiceIf. GeneratedServiceIf is what calls the next method, which is stored in MethodInfo. MethodInfo defaults to a no-op function for now. For ConsensusService, it's ConsensusService, which is generated by a custom protobuf script that links methods in ConsensusService to the properties of MethodInfo, and we'll do that in the next diff.

Test Plan: The hook currently does nothing. I plan to test this e2e once I connect this all the way to RaftConsensus.

Reviewed By: li-chi

Differential Revision: D53744061

fbshipit-source-id: 1158087b10970d4c727a90205ec0e69e618d5a9a