Fix the stamping of RPC tokens for error responses
Summary:
There are two types of errors when we do UpdateReplica.

The first type sets an error in the response but returns OK, like https://fburl.com/code/71ej3pce.

The second type directly returns a non-ok status: https://fburl.com/code/kydujmo7

In the second case we clear the response before rebuilding a new error response. In this process we lose the RPC token. The leader then reads the error as a token mismatch instead of whatever error occurred.

This almost never causes us trouble, because we pretty much always just reduce the rpc frequency and then retry the append till it succeeds. However, if we want the server to act on specific errors during the append differently (like handle corruption), then we need the original error.

Always stamp the error response with the token by using a guard. If the token is wrong the leader will still reject the response. If it's correct the leader should now see the error code properly.

Test Plan: Inject an early failure in UpdateReplica. The leader should be able to see the proper error now

Reviewed By: alanliang

Differential Revision: D55278588

fbshipit-source-id: 64ee4ef73c96b3ea890ae9f0836ecac1400e4c18