Report corruption to leader on a RPC response
Summary:
If we're not able to append anything at all, report the error as Corruption instead of illegal state. This allows us to put in place mitigations specifically for corruptions on the leader.

For example, the leader can try dropping the log cache and reading from the backing log.

Test Plan: Inject a corruption in the decompression path, the message about corruption should show up on the leader.

Reviewed By: alanliang

Differential Revision: D55276685

fbshipit-source-id: fac76896361f7f4f5e146d874a4884284552878f