Fix problem with preceding id when buffering RPCs
Summary:
When we try to append to a buffer, we're currently setting the preceding OpId to the OpId of the first transaction if the buffer is empty. This is wrong, because the preceding OpId should be the OpId of the transaction before the first transaction.

When we did reads from the log cache this was all fine, because we get the OpId from the cache. With appends it becomes tricky because we don't know the term of the previous transaction without going to log cache.

Carry the preceding_id forward when we reset the buffer. After reset, we can either append or read from log cache. If we append, the last_buffered index either matches (then preceding_id is already correct), or it doesn't, in which we reject the append. If we read from log cache, we reset the preceding_opid based on the read. Either way it'll be "consistent" with the buffer.

Test Plan: I was seeing frequent "corruption" messages with buffering turned on, complaining about mismatched indexes. Start a ring with this patch, the messages should go away.

Reviewed By: alanliang

Differential Revision: D55457478

fbshipit-source-id: 5babb741e57766ed1338283524edeb81a12d3bf2