Enable kudu consensus/ gtests
Summary:
Buckify the kudu unit tests in consensus

These all pass, except the disabled ones.

Had to disable the log tests; they will likely require some surgery to revive.

I also had to disable the raft_consensus_quorum-test, which is a good test but
hangs on every test right now, which means there is a basic setup problem. I
debugged it a little and determined that for some reason the RaftConsensus peer
objects are not calling the ConsensusRound replicated callback after we call
Replicate() on them. Something minor but important changed and I'm not sure
what it is.

Test Plan: `buck test //kudu/consensus/...`

Reviewed By: yichenshen

Differential Revision: D56460606

fbshipit-source-id: 5a979d40c75c65eb14fe9d2e9728c180173f00d1