Enable kudu util/ gtests
Summary:
Buckify the kudu unit tests in util

The tests not disabled currently pass. Some are commented out in the TARGETS
file because they don't compile for some reason that I couldn't figure out in a
short period of time. Some are prefixed with DISABLED_ because they compile
but they fail when run.

Test Plan: `buck test //kudu/util/...`

Reviewed By: yichenshen

Differential Revision: D56455916

fbshipit-source-id: 8bd09670172ad7a7fff3c3266fdb06635b4cba06