Kill connection after max number of timeouts reached
Summary:
Currently, when a client RPC times out, we will re-use the connection and try indefinitely. The connection is in some stale state and only a reconnection will revive the situation.

This diff introduces a feature where a connection will be destroyed after X amount of consecutive timeouts. Once the connection is destroyed, the next RPC will re-create the connection.

Test Plan: Added integration tests

Reviewed By: yichenshen

Differential Revision: D56921832

fbshipit-source-id: 05f3514b96b2f940b93187f671817accafeb17f0