Remove SASL dependency from plugin + tests
Summary:
If we have use_normal_tls = 2 today (and we do) and certs are set, kudu will never need to load the SASL lib, so we don't need this.

Set the flags default for normal_tls to match use_normal_tls = 2 to make this the case for tests. However, tests don't have prod certs, so we need them to accept the self signed certs as well.

Test Plan:
Run the ring tests. Because SASL path is unset they are already unable to load SASL, whether it's there or not.

If we don't set the flag to skip TLS verify these tests will fail as TLS connections will fail. With the flag set tests should pass and we can see nodes talking to each other.

Reviewed By: alanliang

Differential Revision: D57083314

fbshipit-source-id: 90aee38422f7d9d8661648f991e0dae23fd29c67