Prevent overflowing quorum calculation
Summary:
We compute the commit watermark by using a sorted list of commit quorum instances.

We then count from the back to get the watermark.

Under normal circumstances quorum size is always >=1, so this will land on one of the instances in the list.

However, quorum-fixer can force this to be 0, so this array will overflow.

Reset it to take the last element in this case. Any write will always advance the writer's own index, so taking the last index is correct here (commit size == 0 means ignore commit quorum anyway).

Test Plan: Added a new unittest. This test would have crashed before this fix in debug mode. It passes now.

Reviewed By: alanliang

Differential Revision: D57885244

fbshipit-source-id: 054ebd7eeabb512819c9f5624321a91fb7e6d55e