Remove unused code in fbcode/kudu
Summary:
This commit removes references to, and code bounded by, the flag ``FB_DO_NOT_REMOVE``.
This flag was used for easy rebasing on the upstream version of kuduraft. The flag, and its bounded code, are no longer needed in this codebase.

Test Plan: Tests passed with ``buck2 test fbcode//kudu/...``.

Reviewed By: yichenshen

Differential Revision: D57698315

fbshipit-source-id: 255e0174b8068a68fd78dd0429c7da5b837782ea