Revive kudu tools action common
Summary:
We didn't compile this since nothing was actually using it. But reviving it now because this thing is super useful for firing RPCs to Raft.

The BuildProxy method lets you create a client to talk to Raft protobuf. There's a bunch of places we can use this in.

Test Plan: Tools common action should build

Reviewed By: alanliang

Differential Revision: D58907298

fbshipit-source-id: c462a3624fd89d47c2ca40b64f3a5181106bbf35