Remove rpl_raft_wait_before_using_voting_history_secs
Summary:
This was supposed to wait for a couple of seconds before using voter history.

However, the sync nature of votes basically means this is only every exercised when the vote itself comes in after the specified time, which is not really the expected behaviour.

We've stablized voter history and already turned this off [everywhere in prod](https://fburl.com/code/86ocr7tf). Remove the option.

Test Plan: unittests

Reviewed By: li-chi

Differential Revision: D59650354

fbshipit-source-id: a7ef9524ceb84fe66cf78ad16c1190f2cd9f6a64