Populate the consideredQuorums field for election decision and use it for logging
Summary:
Currently logging is guessing the regions based on the mechanism, and does not support voter history at all.

With this change, the quorums will be populated directly from the vote counter, so we can be sure what quorums are used and which ones are not.

Test Plan: Upload to merlin and run some elections, we should be able to see the vote tally reflect the regions.

Reviewed By: li-chi

Differential Revision: D61644649

fbshipit-source-id: cf88ae5307b6ebb2b47c09e0fdde02c9c3d2e02d