Add last pruned term to self vote
Summary:
Last pruned term refers to the last term that we garbage collected.

Voter history uses this to determine if a vote history is impossible to get. So if a term is pruned, voter history takes the most conservative approach and just assumes that the instance might have voted for a potential leader.

In our initial vote, we forgot to populate the last pruned term, so it defaults to the election term. This causes voter history to be slightly too conservative when it comes to the self vote. We have the last pruned term in cmeta, just fill it in.

For reference, refer to RaftConsensus::FillVoteResponsePreviousVoteHistory for how we do it for normal votes

Test Plan: ring tests

Reviewed By: li-chi

Differential Revision: D61979471

fbshipit-source-id: a3e127a94b4cebf240bd42567c0e3731c36b52af