Change the bool pair for decision into an Enum
Summary:
Our decision is a pair of bools today. That gives 4 states. (not satisfied + not possible), (not satisfied + possible), (satisfied, not possible), (satisfied, possible).

The third case is actually invalid since statisfied on it's own means it's possible. We have a lot of logic around dealing with 4 states when only 3 are valid, and things get complicated.

Use a enum with 3 states instead so things are clearer. It's either WON, LOST, or UNDECIDED.

Test Plan: This is a refactor, no actual logic changes. Run the flexible vote tests as well as the ring tests.

Reviewed By: li-chi

Differential Revision: D61690760

fbshipit-source-id: 32c46718d516f94c03d88068acdbf8571d7ebae0