Invert the loop functions
Summary:
We have a function that checks for majority for a list of quorums, and another that checks for majority in on quorum.

Makes more sense to have the list function call the single function, instead of the single function making a list of one and calling the list function.

Test Plan: Only refactoring, run leader_election-tests and flexible voter tests.

Reviewed By: li-chi

Differential Revision: D61748887

fbshipit-source-id: 8cd03e4c315091ea6197a1fc77004fc442178d4f