Change to Log::ReadReplicatesInRange to return ref counted objects rather than raw pointers
Summary: We will be integrating warm storage reads at the BinlogWrapperBase level and it would be difficult if we have to work with raw pointers. One problem is that, if for some reason we had to read the same range of Replicates during a retry, we would have to restart the underlying Warm Storage stream which is quite expensive to initialize. If the interface returned scoped ref counted objects (similar to shared pointers), we would be able to cache the Replicates and avoid re-opening the stream if we had to retry.

Test Plan: Existing tests

Reviewed By: abhinav04sharma

Differential Revision: D61818066

fbshipit-source-id: 88956cc55ae61713ea2747a882f7292639d32129