Do not destroy connections directly
Summary:
Connections should be not be destroyed directly if we time out on stuff. It could still be in the middle of doing something else and those needs cleanup.

Instead we should mark it for shutdown. This prevents new calls from using the connection and when it's ready, it'll be shut down.

The motivation was S438943. From code reading, I suspect if we tear the connection down directly, we'll get stuck on negotiation, since during negotiation the socket is a blocking socket, and won't close till data is read. In any case, it probably best not to just yank out the connection while stuff is still active.

Test Plan: There are integration tests for this in kudu/rpc/rpc-test.cc, run them.

Reviewed By: alanliang

Differential Revision: D62672395

fbshipit-source-id: 6a079e4a8e885beb62b6686ed11a07f0bfdcdbca