Avoid unnecessarily creating new warm storage streams during catchup
Summary:
When we send transactions 50 - 100 to a replica, it might only append 50-75, leaving some on the table. Next time, it may request 76 - 125.

This was handled fine with the CatchupStream. However, with compression, we may ask the stream to return max_size_bytes, say 100 mb of transactions. The stream may return 50-100 which is 101 mb. We then compress these transactions and the total size becomes 90mb. Since we have 10mb left, we will request the stream for the next 10 mb worth of transactions, which happens to be 101-105. Now as mentioned before, the replica may not append all transactions. If it had appended up to 75, the next call to retrieve 76 onwards will require a new stream since the cache only contains the last requests worth of transactions which was 101-105.

There are many ways to handle this but the easiest for now is to request a little more than asked so that don't run into a situation where we make second round to fetch more before returning to the client.

I believe this is a safe and quick compromise for now. This does deserve a bigger refactor to make things more simpler.

Test Plan: Deployed to stage-1 and ensured we were able to catchup without creating new streams.

Reviewed By: abhinav04sharma

Differential Revision: D63078828

fbshipit-source-id: 9c8703464aad36b180f6bd9892b765a1a03aefac