Add new routing policy to support region group
Summary:
This diff adds a new region group based routing policy for myraft to reduce [x-atlanic network cost](https://docs.google.com/document/d/17Zra8dY5INtNagYvVqGOE6ShpcOfneY9ETaomCAtW6w/edit).

The idea is for replicas in EU which is far from US, it should select one replica as proxy for the rest of replicas to avoid sending cross ocean log messages multiple times. It is to achieve that by the abstraction of region group, all replicas in EU can be put into one group and replicas in that group can share one proxy. With the region group support, if the leader is in EU, the replicas in US can be put into one group and they can use one replica as proxy to avoid x-ocean traffic as well.

Note: the region group proxy will only be proxy for database replicas, LBU will still use the database replica in the same region as proxy. For example, assume a replica set has the following setup
- prn-db1 (leader), prn-lbu1, prn-lbu2
- frc-db1, frc-lbu1, frc-lbu2
- atn-db1, atn-lbu1, atn-lbu2
- lla-db1, lla-lbu1, lla-lbu2
- cln-db1, cln-db2, cln-lbu1, cln-lbu2, cln-lbu3, cln-lbu4

The routing topology for all the replicas will be like this
- prn-db1  -> (prn-lbu1, prn-lbu2, frc-db1, atn-db1, cln-db1) # cln will be used as the proxy for region group of cln and lla since it is closer to leader at prn
- frc-db1 -> (frc-lbu1, frc-lbu2)
- atn-db1 -> (atn-lbu1, atn-lbu2)
- cln-db1 -> (lla-db1, cln-db2, cln-lbu1, cln-lbu2, cln-lbu3, cln-lbu4)
- lla-db1 -> (lla-lbu1, lla-lbu2)

This will slightly increase the latency for message reaches to replicas in the region group being proxy-ed because they need to wait until the proxy gets the message then get the message from the proxy peer.

Test Plan: see the testing in the following diffs

Reviewed By: yichenshen

Differential Revision: D61418296

fbshipit-source-id: 555efc3bdfabc74eccce7553a4dc07c6920a1377