Fix BLS crash and some refactor for StateMachineMetricsInterface
Summary:
BLS was returning nullptr for this, so when it receives an update it triggers a nullptr exception.

Guard the block with a null check first.

Also split out the interface, added some docs, and initialized it earlier on via server opts. This helps prevent early lifecycle issues (i.e. for mysql we can expect the ptr to always be valid once running).

Test Plan: Canary BLS. Previously it was crash looping. This should solve it.

Reviewed By: li-chi

Differential Revision: D66407800

fbshipit-source-id: e29a9804f565d1d5df51e5fb2d60904a369e07f2