fix off-by-one error for vote history
Summary:
This diff is to fix an off-by-one bug in vote history logic. For the given case:
- prn_lbu_1, vote for atn_lbu_1 on term 330, last prune term 328
- prn_lbu_2, last prune term 330
- prn_db_1, last prune term 328

If prn_db_1 wants to elect itself as leader on term 332, it should assume atn_lbu_1 might have won election and include atn in the extend vote regions and tries to get vote from atn replicas. The old code has off-by-one issue where it doesn't treat last prune term 330 as the case it might vote for atn_lbu_1 so that for term 330, it assumes there is no majority for the vote and skip checking atn replicas. This diff is to fix that by counting last prune term 330 as unknown vote and force candidate to check atn replicas for vote history as well.

Test Plan: unit test

Reviewed By: yichenshen

Differential Revision: D67184901

fbshipit-source-id: 08ef7530e7b913353dcf87914a0ddff9996ab15a