Abandon connections that are aborted after negotiation
Summary:
Connection negotiation happens on another thread pool. The race here is that we shutdown the connection, then the negotiation completes and we schedule the post negotiation task.

The post negotiation task will be the sole holder of the connection. It reconnects the connection to the evb, but immeidately lets go of it's reference and the connection destructs with the evb still attached.

Bascially, Reactor thread doesn't have control over the connection during negotiation. As such, we need to mark a mark on the connection if it's not under control and only shutdown after we regain control.

The flow is:

New connection
```
1. Get inbound connection
2. Start inbound negotiation async
3. Mark connection as running negotiation
4. Add to connection list
```

After negotiation, schedule back on reactor
```
1. Check if connection should continue bootstrapping and shutdown if not
```

Reset
```
1. Go through connection list
2. If connection is negotiating, mark that connection should shutdown
3. If connection is no negotiating just shutdown.
```

Each group is a reactor task and thus will be exclusive to each other. As such, a connection in the list is either negotiating or negotiated. We shutdown the former after negotiation. We shutdown the later immediately.

Test Plan:
Sanity check the process by running a Raft ring and ensuring that connections are sound.

Added a test for it. Without the change the test will crash on a assert

Reviewed By: alanliang

Differential Revision: D71436522

fbshipit-source-id: db55f27972687c409a6ce6a506d8cf607a762a49