Fix deprecated unittest asserts
Summary:
In Python 3.2+, a number of asserts were deprecated in the unittest module: https://docs.python.org/3/library/unittest.html#deprecated-aliases.

In Python 3.12, these asserts were deleted completely.
The files in this diff still use the deprecated asserts.

This codemod attempts to fix uses of the deprecated asserts, replacing them with non-deprecated equivalents. It is a naive codemod, using sed.

Differential Revision: D72616297