Log Cache should read up to min_pinned_op_index_ instead of next_sequential_op_index_
Summary:
LogCache has a bug today where we bound the reads up to next_sequential_op_index_ but the transaction for next_sequential_op_index_ may exist on cache but might not have written to disk yet.

During ingestion testing with skipping local cache (to avoid overhead of uncompressing payloads from cache) and going to disk, I get into a situation where we try to read from disk up to next_sequential_op_index_, but it may not make it there and we get a FileNotFound/OpIdNotFound.

To avoid this, we should read up to min_pinned_op_index_-1, which is index for last appended op index

Test Plan: Tested in stage-1.

Reviewed By: abhinav04sharma

Differential Revision: D75153887

fbshipit-source-id: 9abe26861bdf0b63b4a0b105490dc493bc7cc933