Clean up unused flags
Summary:
These are flags that are just left as default everywhere and are never used. Clean up forks in the code by removing them.

- abort_trxs_on_replication_done: always off
- synchronous_transfer_leadership: always on
- region: Not used anywhere
- transfer_timeout_multiple: always 3
- diag_thread_stats: always true
- flush_binlog_per_trx: always true and we always need to flush
- update_lkl_after_new_term_append: causes dataloss if off
- filter_out_bad_quorums_in_lmp: always true, no reason not to

This is not a complete list, it's just a bunch of low hanging fruits.

Test Plan:
Ring and unittests

Each variable I also did zbgs to make sure nothing else is using it

Reviewed By: li-chi

Differential Revision: D75197895

fbshipit-source-id: 625f5624c2165568773dac3dde2173b9ce6a7296