Use targeted transfer instead of untargeted transfer
Summary:
Later on, we'll prevent instances undergoing PLS from starting elections. However, if we do untargeted elections, we'll fail to transfer leadership if the selected target also has PLS active, and simply step down.

Instead, we need to run mock elections to be able to know if the target is viable. This way, we can fail after and gracefully proceed to try the next instance.

Test Plan:
Upload the RPM and run a PLS. Make sure promotion goes through.

Rollback Plan:

Differential Revision: D79297253

fbshipit-source-id: cc2b888f8523fe6c11f6dac4b4ff1f0370eef98d