Do not unlock peer till we managed to get the proxy
Summary:
I see a crash on the LOG fatal line. When examining the core dump I see that the peer is closed.

This can happen when while we're preparing the RPC a promotion happens. On promotion we'll reset the peer manager and close the peers since we don't need to be sending RPCs anymore.

The close happens under the peer lock and will clear the proxy map. As you can see, the close can sneak in between `l.unlock()` and `peer_proxy_pool->Get()`. I confirmed this through a coredump by examining `this->closed_ == true`. This is inited to false and only flipped during `Close()`.

Instead, we should just get the peer proxy under the lock. This is just a map lookup and should be relatively quick.

Test Plan: Ring tests and MTRs to make sure this doesn't cause any problems.

Reviewed By: mpercy, alanliang

Differential Revision: D84640706

fbshipit-source-id: 82df6a793776528ab4e56e17e0db959dfd3eaf4e