Rename ThreadPool classes to KuduThreadPool
Summary:
Change the current ThreadPool class to KuduThreadPool. Then reduce ThreadPool to an interface.

This allows us to keep the syntax almost everywhere except for the constructor callsites. These are done with ThreadPoolBuilder so we extract that from ThreadPool and make it it's own BUCK target.

From there, we can insert a switch into ThreadPoolBuilder and we can substitute in other ThreadPool classes.

Test Plan: Build and existing tests

Differential Revision: D88762270

fbshipit-source-id: 50b67698ebfaef58922e495d24b7ec2241154ead