Change peer proxy pool to ConcurrentHashMap
Summary:
This stores the peers to send requests to. It's currently just a rwlock around an unordered map. The whole thing only provides Get, Put and Clear(). It stores and returns shared_pointers.

This should be perfect for a folly::ConcurrentHashMap since the same guarantees would apply. Peers don't churn very much and there aren't a lot of them, so the delayed destruction shouldn't matter that much.

Test Plan: Ring tests and MTRs.

Differential Revision: D89314343

fbshipit-source-id: 219318934ab6cc86c2cc41d1f4d81ca04181c603