Fix use-after-free in FunctionGaugeDetacher by capturing shared_ptr to it.
Summary: Metric Entity tracks function guages but expects another shared_ptr to be kept to it outside of its map as sole ref being with it is taken as sign of the metric no longer being needed. Some metrics are embedding a shared_ptr directly in their class, while others are relying on the detacher capturing a reference. This was done by refcnt system implicitly and needs to be done explicitly by now in detacher callbacks.

Test Plan:
- Built `fbcode//kudu/util:common` successfully
- Ran `fbcode//kudu/util:metrics-test` - 16 tests passed
- Ran `fbcode//kudu/util:mt-metrics-test` - 2 tests passed

Reviewed By: yichenshen, alanliang

Differential Revision: D89350150

fbshipit-source-id: f749fe2aba9579c543e7fff3800710fabbcec1da